Separate types into separate resource files
