+++ .gitignore
@@ -12,3 +12,4 @@ node_modules/
 /typedoc/back-end/
 /typedoc/shared/
 /.rpt2_cache/
+.vscode/
