+++ src/back-end/lib/hooks/logger.ts
@@ -1,6 +1,6 @@
 import { RouteHook } from 'back-end/lib/server';
 import chalk from 'chalk';
-import { Session } from 'shared/lib/types';
+import { Session } from 'shared/lib/resources/session';
 
 const hook: RouteHook<unknown, unknown, number, Session>  = {
 
@@ -13,6 +13,6 @@ const hook: RouteHook<unknown, unknown, number, Session>  = {
     request.logger.info(`${chalk.gray('<-')} ${response.code} ${Date.now() - startTime}ms`);
   }
 
-}
+};
 
 export default hook;
