+++ src/shared/lib/resources/affiliation.ts
@@ -1,0 +1,5 @@
+export enum MembershipType {
+    Owner = 'OWNER',
+    Member = 'MEMBER',
+    Pending = 'PENDING'
+}
