+++ src/shared/lib/types.ts
@@ -28,39 +28,3 @@ export enum ClientHttpMethod {
   Delete = 'DELETE',
   Options = 'OPTIONS'
 }
-
-export enum UserType {
-  Vendor = 'VENDOR',
-  Government = 'GOV',
-  Admin = 'ADMIN'
-}
-
-export enum UserStatus {
-  Active = 'ACTIVE',
-  InactiveByUser = 'INACTIVE_USER',
-  InactiveByAdmin = 'INACTIVE_ADMIN'
-}
-
-export enum MembershipType {
-  Owner = 'OWNER',
-  Member = 'MEMBER',
-  Pending = 'PENDING'
-}
-
-export interface Session {
-  id: Id;
-  token?: string;
-  user?: User;
-}
-
-export interface User {
-  id: Id;
-  type: UserType;
-  status: UserStatus;
-  name: string;
-  email?: string;
-  avatarImageUrl?: string;
-  notificationsOn: boolean;
-  acceptedTerms: boolean;
-  idpUsername: string;
-}
